src/rag/requirements/basic-requirements.txt
requests
sentence_transformers
llama-index-retrievers-bm25
langchain_core
langchain_chroma
langchain_ollama
chromadb
fastapi
langchain_text_splitters
langchain
langchain_community
langchain_experimental
neo4j
json_repair
grandalf
python-dotenv
llama-index-llms-langchain
llama-index-llms-ollama
llama-index-graph-stores-neo4j
llama-index-llms-openai